requirements.txt
blinker==1.6.2
boto3==1.26.91
botocore==1.29.91
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Flask==2.3.2
fonttools==4.39.0
idna==3.4
imgviz==1.7.2
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
labelme==5.2.1
MarkupSafe==2.1.2
matplotlib==3.7.1
natsort==8.3.1
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
psutil==5.9.4
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
QtPy==2.3.1
requests==2.28.2
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.16.0
six==1.16.0
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
typing-extensions==4.5.0
ultralytics==8.0.53
urllib3==1.26.15
Werkzeug==2.3.4
zipp==3.15.0
